requirements.txt
Django>=5.2,<7.0
djangorestframework>=3.16,<4.0
psycopg[binary]>=3.2,<4.0

